- [x] table.types.ts and the types in TableRowActionsRuntime need to be consolidated to the main types file.
