+++ src/components/table-quadrone/TidyInlineActivitiesList.svelte
@@ -8,23 +8,20 @@
   import { Activities } from 'src/features/activities/activities';
   import type {
     ActivityItemContext,
+    ActivityTableAction,
+    ActivityTableActionData,
     ActorSheetQuadroneContext,
   } from 'src/types/types';
   import { SheetSections } from 'src/features/sections/SheetSections';
   import TableRowActions from 'src/components/table-quadrone/parts/TableRowActions.svelte';
   import { getSheetContext } from 'src/sheets/sheet-context.svelte';
-  import TableRowActionsRuntime, {
-    type ActivityTableAction,
-    type ActivityTableActionData,
-  } from 'src/runtime/tables/TableRowActionsRuntime.svelte';
+  import TableRowActionsRuntime from 'src/runtime/tables/TableRowActionsRuntime.svelte';
   import TidyTableHeaderRow from './TidyTableHeaderRow.svelte';
   import TidyTableHeaderCell from './TidyTableHeaderCell.svelte';
   import { ThemeQuadrone } from 'src/theme/theme-quadrone.svelte';
   import { ActivityColumnRuntime } from 'src/runtime/tables/ActivityColumnRuntime.svelte';
   import TidyTableCustomHeaderCells from './parts/TidyTableCustomHeaderCells.svelte';
   import TidyTableCustomCells from './parts/TidyTableCustomCells.svelte';
-  import type { TidyTableAction } from './table-buttons/table.types';
-  import type { Activity5e } from 'src/foundry/dnd5e.types';
 
   interface Props {
     item?: Item5e | null;
