--- src/components/table-quadrone/table-buttons/table.types.ts
@@ -1,14 +1,0 @@
-import type { Component, ComponentProps } from 'svelte';
-
-export type TidyTableActionArgs<TData extends object> = {
-  data: TData;
-};
-
-export type TidyTableAction<
-  TComponent extends Component<any>,
-  TData extends object,
-> = {
-  component: TComponent;
-  props: (args: TidyTableActionArgs<TData>) => ComponentProps<TComponent>;
-  condition?: (args: TidyTableActionArgs<TData>) => boolean;
-};
