+++ src/features/activities/activities.ts
@@ -1,17 +1,8 @@
 import type { Activity5e } from 'src/foundry/dnd5e.types';
 import { TidyHooks } from 'src/foundry/TidyHooks';
 import type { Item5e } from 'src/types/item.types';
-import type {
-  ActivityItemContext,
-  Actor5e,
-  CharacterItemContext,
-  LinkedUses,
-  NpcItemContext,
-} from 'src/types/types';
-import TableRowActionsRuntime, {
-  type ActivityTableAction,
-  type ActivityTableActionData,
-} from 'src/runtime/tables/TableRowActionsRuntime.svelte';
+import type { ActivityItemContext, LinkedUses } from 'src/types/types';
+import TableRowActionsRuntime from 'src/runtime/tables/TableRowActionsRuntime.svelte';
 
 export class Activities {
   static isConfigurable(activity: Activity5e) {
