+++ src/todo-columns-loadout-updates.md
@@ -8,7 +8,6 @@
 - [ ] // TODO: This happens twice. Where should this data prep go?
 - [ ] Refactoring consideration: Row Actions don't need both presence/absence as well as an optional condition callback. We also don't need to pull row actions until we have an item to be evaluated. Consider eliminating presence/absence and just evaluating conditions each time an item needs row actions. `TableRowActionsRuntime.getEncounterMemberRowActions(context: ActorSheetQuadroneContext)` and others like it do not need the full context object in order to return row actions, just some intel like editable and unlocked, plus the document entry.
 - [ ] Extract a TidyRowActionsCell? Determine if this is feasible, given other row action component updates.
-- [ ] table.types.ts and the types in TableRowActionsRuntime need to be consolidated to the main types file.
 
 ## Stretch
